Keep coverage overlay when no new tests found
When Jest is run for only changed files the tests results that follow
the "No tests found related to files changed since the last commit"
message are empty. This causes the coverage overlay to be cleared.

This commit skips the coverage overlay update when the test results
follow:
* the "no tests found..." message
* the "no tests found..." then "watch usage..." messages
<br/>

This commit depends on a change to `jest-editor-support`. We
shouldn't merge this until facebook/jest#5208 is good to go and we
update `package.json` accordingly.