test-provider: adjust workspace location handling
The VSCode reported workspace path is unresolved
(microsoft/vscode#18837).  When the repository is hosted in a path
substituted path (e.g. `subst X: C:`), the test path will not be
properly translated as the drive is different.  This may potentially
also impact paths which are hosted within a NT junction.  Explicitly
resolve the workspace path to match the fact that `jest` will report
paths which have been resolved to the real path.